Add functionality that meets user story 1 requirements
This change also contains tests for user story 1.

User story 1 was unclear as to what would happen once the account numbers were
parsed, so I went ahead and droppped them into a file that mirrors the input
filename.

The general approach to parsing the numbers was to look at the input values
3 characters at a time and map those values to a prime. Since each number is
composed of 9 total characters, this looks like:

 _     2 3 5
|_| -> 7 11 13
|_|    17 19 23

When a correct character is encountered int he correct location, we multiply
that lines value by the corresponding positions prime. Once we have produced
the three line values for each number, we multiply these values together to
give us the number we will use to look up the expected value. If we get a hit
on our product from this process, the input was valid for this number, if we do
not it was malformed in some way.